+++ HISTORY.rst
@@ -21,6 +21,8 @@ latest (unreleased)
 
 **Features and Improvements**
 
+* ALCYN-1854: Make the delivery carrier on sale order required and set its default value to Alcyon Delivery Method
+
 **Bugfixes**
 
 **Build**